Conflicts:
	Makefile

+++ installer/mk-install-iso.sh
@@ -11,10 +11,8 @@
 #	End
 #
 
-set -o errexit		# exit if error...insurance ;)
-set -o nounset		# exit if variable not initalized
-set +h			# disable hashall
-PRGNAME=${0##*/}	# script name minus the path
+set +x                 # disable hashall
+PRGNAME=${0##*/}	    # script name minus the path
 source config.inc		#	configuration parameters
 source function.inc		#	commonn functions
 LOGFILE=/var/log/"${PRGNAME}-${LOGFILE}"	#	set log file name
@@ -22,40 +20,40 @@ LOGFILE=/var/log/"${PRGNAME}-${LOGFILE}"	#	set log file name
 
 # Grab the name of the iso file 
 if [ $# -lt 2 ]
-	then
-		echo "Usage : " $0 " <output-iso-with-path>  <tools path>"
-		exit 1
+then
+    echo "Usage : " $0 " <output-iso-with-path>  <rpms-path> <pkg-list-path>"
+    exit 1
 fi
 ISO_OUTPUT_NAME=$1
-TOOLS_PATH=$2
+RPMS_PATH=$2
 PACKAGE_LIST_FILE=$3
-
-
+RPM_LIST=$4
+STAGE_PATH=$5
+ADDITIONAL_FILES_TO_COPY_FROM_STAGE=$6
+LIVE_CD=$7
+OUTPUT_DATA_PATH=$8
+PHOTON_COMMON_DIR=$(dirname "${PACKAGE_LIST_FILE}")
+PACKAGE_LIST_FILE_BASE_NAME=$(basename "${PACKAGE_LIST_FILE}")
 #- Step 3 Setting up the boot loader
+WORKINGDIR=${BUILDROOT}
+BUILDROOT=${BUILDROOT}/photon-chroot
 
-mkdir ${BUILDROOT}/isolinux
-cp BUILD_DVD/isolinux/* ${BUILDROOT}/isolinux/
+cp -r BUILD_DVD/isolinux ${WORKINGDIR}/
 
-mv ${BUILDROOT}/boot/* ${BUILDROOT}/isolinux/
-rmdir ${BUILDROOT}/boot
-ln -s isolinux ${BUILDROOT}/boot
+if [ "$LIVE_CD" = true ] ; then
+    mv ${WORKINGDIR}/isolinux/live-menu.cfg ${WORKINGDIR}/isolinux/menu.cfg
+fi
 
-find ${BUILDROOT} -name linux-[0-9]*.rpm | head -1 | xargs rpm2cpio | cpio -iv --to-stdout ./boot/vmlinuz* > ${BUILDROOT}/boot/vmlinuz
+cp sample_ks.cfg ${WORKINGDIR}/isolinux/
 
-cp -r ../installer ${BUILDROOT}/${PARENT}
-# replace default package_list with specific one
-cp $PACKAGE_LIST_FILE ${BUILDROOT}/${PARENT}/installer/package_list.json
-rm -f ${BUILDROOT}/${PARENT}/installer/*.pyc
+find ${BUILDROOT} -name linux-[0-9]*.rpm | head -1 | xargs rpm2cpio | cpio -iv --to-stdout ./boot/vmlinuz* > ${WORKINGDIR}/isolinux/vmlinuz
 
+rm -f ${BUILDROOT}/installer/*.pyc
+rm -rf ${BUILDROOT}/installer/BUILD_DVD
+# Copy package list json files, dereference symlinks
+cp -rf -L $OUTPUT_DATA_PATH/*.json ${BUILDROOT}/installer/
 #ID in the initrd.gz now is PHOTON_VMWARE_CD . This is how we recognize that the cd is actually ours. touch this file there.
-touch ${BUILDROOT}/PHOTON_VMWARE_CD
-
-#- Step 4 - Move /etc /var /root /home to /fake/needwrite
-
-mkdir -p ${BUILDROOT}/fake/{needwrite,ramdisk}
-mv ${BUILDROOT}/var/ ${BUILDROOT}/tmp ${BUILDROOT}/fake/needwrite/
-ln -s fake/needwrite/var ${BUILDROOT}/var
-ln -s fake/needwrite/tmp ${BUILDROOT}/tmp
+touch ${WORKINGDIR}/PHOTON_VMWARE_CD
 
 # Step 4.5 Create necessary devices
 mkfifo ${BUILDROOT}/dev/initctl
@@ -68,50 +66,102 @@ mknod ${BUILDROOT}/dev/sda b 8 0
 
 #- Step 5 - Creating the boot script
 mkdir -p ${BUILDROOT}/etc/systemd/scripts
-cp BUILD_DVD/create_ramdisk ${BUILDROOT}/bin/create_ramdisk
 
 # Step 6 create fstab
 
 cp BUILD_DVD/fstab ${BUILDROOT}/etc/fstab
 
 #- Step 7 - Create installer script
+if [ "$LIVE_CD" = false ] ; then
 
 cat >> ${BUILDROOT}/bin/bootphotoninstaller << EOF
 #!/bin/bash
-create_ramdisk
-cd ${PARENT}/installer
-./isoInstaller.py 2> /var/log/installer && shutdown -r now
+cd /installer
+./isoInstaller.py --json-file=$PACKAGE_LIST_FILE_BASE_NAME 2> /var/log/installer && shutdown -r now
 /bin/bash
 EOF
 
 chmod 755 ${BUILDROOT}/bin/bootphotoninstaller
 
+fi
+
+cat >> ${BUILDROOT}/init << EOF
+mount -t proc proc /proc
+/lib/systemd/systemd
+EOF
+chmod 755 ${BUILDROOT}/init
+
 #adding autologin to the root user
 sed -i "s/ExecStart.*/ExecStart=-\/sbin\/agetty --autologin root --noclear %I $TERM/g" ${BUILDROOT}/lib/systemd/system/getty@.service
 
-sed -i "s/root:.*/root:x:0:0:root:\/root:\/bin\/bootphotoninstaller/g" ${BUILDROOT}/etc/passwd
+#- Step 7 - Create installer script
+if [ "$LIVE_CD" = false ] ; then
 
-mkdir -p ${BUILDROOT}/mnt/photon-root
-cp $TOOLS_PATH/tools.tar.gz ${BUILDROOT}/${PARENT}/
+    sed -i "s/root:.*/root:x:0:0:root:\/root:\/bin\/bootphotoninstaller/g" ${BUILDROOT}/etc/passwd
+
+fi
+
+mkdir -p ${BUILDROOT}/mnt/photon-root/photon-chroot
+rm -rf ${BUILDROOT}/RPMS
+
+#cp -r ${RPMS_PATH} ${WORKINGDIR}/
+(
+cd ${RPMS_PATH}
+mkdir ${WORKINGDIR}/RPMS
+for rpm_name in $RPM_LIST; do
+    FILENAME="`find . -name "$rpm_name-[0-9]*" -type f`"
+    if [ -n "$FILENAME" ]; then
+        cp --parent $FILENAME ${WORKINGDIR}/RPMS/;
+    fi
+done
+)
+
+# Work in sub-shell using ( ... ) to come back to original folder.
+(
+cd $STAGE_PATH
+for file_name in $ADDITIONAL_FILES_TO_COPY_FROM_STAGE; do
+    if [ -n "$file_name" ]; then
+        cp $file_name ${WORKINGDIR}/;
+    fi
+done
+)
 
 #creating rpm repo in cd..
-createrepo --database ${BUILDROOT}/${PARENT}/RPMS
+createrepo --database ${WORKINGDIR}/RPMS
 
-rm -rf ${BUILDROOT}/${PARENT}/LOGS
+rm -rf ${BUILDROOT}/LOGS
 
-#Remove our rpm database as it fills up the ramdisk
-rm -rf ${BUILDROOT}/var/lib/rpm
-find "$BUILDROOT"/{,usr/}lib -type f -exec strip --strip-debug '{}' ';' > /dev/null 2>&1
-find "$BUILDROOT"/{,usr/}{bin,sbin} -type f -exec strip --strip-unneeded '{}' ';' > /dev/null 2>&1
-rm -rf $BUILDROOT/tools
-#Step 9 Generate the ISO!!!!
+if [ "$LIVE_CD" = false ] ; then
+    # Cleaning up
+    #Remove our rpm database as it fills up the ramdisk
+    rm -rf ${BUILDROOT}/home/*
+    rm -rf ${BUILDROOT}/var/lib/rpm
+
+    # Remove the boot directory
+    rm -rf ${BUILDROOT}/boot
+
+    #Remove the include files.
+    rm -rf ${BUILDROOT}/usr/include
+
+    # TODO: mbassiouny, Find a clean way to do that
+    for i in `ls ${BUILDROOT}/usr/share/`; do
+    	if [ $i != 'terminfo' -a $i != 'cracklib' -a $i != 'grub' ]; then
+    		rm -rf ${BUILDROOT}/usr/share/$i
+    	fi
+    done
+
+fi
+
+# Generate the intird
 pushd $BUILDROOT
+(find . | cpio -o -H newc --quiet | gzip -9 ) > ${WORKINGDIR}/isolinux/initrd.img
+popd
+rm -rf $BUILDROOT
+
+#Step 9 Generate the ISO!!!!
+pushd $WORKINGDIR
 mkisofs -R -l -L -D -b isolinux/isolinux.bin -c isolinux/boot.cat \
 		-no-emul-boot -boot-load-size 4 -boot-info-table -V "PHOTON_$(date +%Y%m%d)" \
-		$BUILDROOT >$ISO_OUTPUT_NAME
+		$WORKINGDIR >$ISO_OUTPUT_NAME
 
 popd
-
-
-
-
